Fix local tests.

+++ local/run-tests-ci-locally.sh
@@ -3,6 +3,13 @@
 # This script is an helper to run some tests from Gitlab-ci config in a local
 # environment with docker-compose.yml
 
+set -o nounset
+set -o errexit
+trap 'echo "Aborting due to errexit on line $LINENO. Exit code: $?" >&2' ERR
+set -o errtrace
+set -o pipefail
+IFS=$'\n\t'
+
 ###############################################################################
 # Environment
 ###############################################################################
@@ -16,18 +23,6 @@ _red=$'\e[1;31m'
 _grn=$'\e[1;32m'
 _blu=$'\e[1;34m'
 _end=$'\e[0m'
-
-# Grab .docker-compose.env files to simulate CI variables.
-if [ -f "local/.docker-compose.env" ]; then
-  source local/.docker-compose.env
-else
-  if [ -f ".docker-compose.env" ]; then
-    source .docker-compose.env
-  else
-    printf "%s[ERROR]%s Missing .docker-compose.env file.\\n" "${_red}" "${_end}"
-    exit 1
-  fi
-fi
 
 _CMD="${1:-"help"}"
 
@@ -88,10 +83,10 @@ HEREDOC
 
 _status() {
   _dkexec /scripts/run-tests.sh
-  docker exec -d ci-drupal /scripts/start-selenium-standalone.sh
+  docker exec -d ci-drupal bash -c "/scripts/start-selenium-standalone.sh"
   sleep 2s
   _dkexec bash -c "curl -s http://localhost:4444/wd/hub/status | jq '.'"
-  docker exec -d ci-drupal bash -c "/usr/bin/chromium --no-sandbox --disable-gpu --headless --remote-debugging-address=0.0.0.0 --remote-debugging-port=9222 --window-size=1920,1080"
+  docker exec -d ci-drupal bash -c "/scripts/start-chrome.sh"
   sleep 1s
   _dkexec bash -c "curl -s http://localhost:9222/json/version | jq '.'"
 }
@@ -100,26 +95,25 @@ _status() {
 _tests_prepare() {
   printf ">>> [NOTICE] tests_prepare\\n"
 
-  _dkexec bash -c "cp -u ${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php ${CI_PROJECT_DIR}"
+  echo _dkexec bash -c "cp -u \${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php \${CI_PROJECT_DIR}"
 
   _dkexec robo prepare:folders
   
   # Apache launch is entrypoint.
-  _dkexec apache2-foreground&
+  # docker exec -d ci-drupal bash -c "apache2-foreground"
 
   # Prepare needed folders, reproduce .test_template
   if [ ${CI_TYPE} == 'custom' ];
   then
-    _dkexec bash -c "cp -u ${CI_PROJECT_DIR}/local/phpunit.local.xml ${CI_PROJECT_DIR}/web/core/phpunit.xml"
+    _dkexec bash -c "cp -u \${CI_PROJECT_DIR}/local/phpunit.local.xml \${CI_PROJECT_DIR}/web/core/phpunit.xml"
   else
-    _dkexec bash -c "cp -u ${CI_PROJECT_DIR}/local/phpunit.local.xml ${WEB_ROOT}/core/phpunit.xml"
+    _dkexec bash -c "cp -u \${CI_PROJECT_DIR}/local/phpunit.local.xml \${WEB_ROOT}/core/phpunit.xml"
   fi
 
   # RoboFile.php is already at root.
-  _dkexec mkdir -p "${BROWSERTEST_OUTPUT_DIRECTORY}/browser_output"
-  _dkexec chmod -R g+s "${BROWSERTEST_OUTPUT_DIRECTORY}"
-  _dkexec chown -R ${APACHE_RUN_USER}:${APACHE_RUN_GROUP} ${BROWSERTEST_OUTPUT_DIRECTORY}
-
+  _dkexec "mkdir -p \${BROWSERTEST_OUTPUT_DIRECTORY}/browser_output"
+  _dkexec "chmod -R g+s \${BROWSERTEST_OUTPUT_DIRECTORY}"
+  _dkexec "chown -R \${APACHE_RUN_USER}:\${APACHE_RUN_GROUP} \${BROWSERTEST_OUTPUT_DIRECTORY}"
 }
 
 # Replicate Build job.
@@ -127,7 +121,7 @@ _build() {
   printf ">>> [NOTICE] build\\n"
   _simulate_cache
 
-  _dkexec bash -c "cp ${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php ${CI_PROJECT_DIR}"
+  _dkexec bash -c "cp \${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php \${CI_PROJECT_DIR}"
 
   _dkexec robo perform:build
 
@@ -153,7 +147,7 @@ _extract_artifacts() {
   if [ -f tmp/artifacts.tgz ]
   then
     mv tmp/artifacts.tgz .
-    _dkexec tar -xzf ${CI_PROJECT_DIR}/artifacts.tgz
+    _dkexec tar -xzf \${CI_PROJECT_DIR}/artifacts.tgz
     mkdir -p tmp
     mv artifacts.tgz tmp/
   else
@@ -165,7 +159,9 @@ _extract_artifacts() {
 _simulate_cache() {
   printf ">>> [NOTICE] simulate_cache\\n"
   _extract_artifacts
-  rm -rf drush scripts composer.json composer.lock .env.example load.environment.php
+  if ! [ ${CI_TYPE} == "project" ]; then
+    rm -rf drush scripts composer.json composer.lock .env.example load.environment.php
+  fi
 }
 
 ####### Tests jobs
@@ -175,7 +171,7 @@ _unit_kernel() {
 
   _build
   _tests_prepare
-  _dkexec robo test:suite "${PHPUNIT_TESTS}unit,${PHPUNIT_TESTS}kernel"
+  _dkexec robo test:suite "\${PHPUNIT_TESTS}unit,\${PHPUNIT_TESTS}kernel"
 }
 
 _code_coverage() {
@@ -183,7 +179,7 @@ _code_coverage() {
 
   _build
   _tests_prepare
-  _dkexec robo test:coverage "${PHPUNIT_TESTS}unit,${PHPUNIT_TESTS}kernel"
+  _dkexec robo test:coverage "\${PHPUNIT_TESTS}unit,\${PHPUNIT_TESTS}kernel"
 }
 
 _functional() {
@@ -192,10 +188,11 @@ _functional() {
   _build
   _tests_prepare
 
-  _dkexec chown -R ${APACHE_RUN_USER}:${APACHE_RUN_GROUP} ${WEB_ROOT}/sites/
-  _dkexec sudo -E -u ${APACHE_RUN_USER} robo test:suite "${PHPUNIT_TESTS}functional"
 
-  _dkexec bash -c "cp -f ${DOC_ROOT}/sites/simpletest/browser_output/*.html ${REPORT_DIR}/functional"
+  _dkexec "chown -R \${APACHE_RUN_USER}:\${APACHE_RUN_GROUP} \${WEB_ROOT}/sites/"
+  _dkexec "sudo -E -u \${APACHE_RUN_USER} robo test:suite \${PHPUNIT_TESTS}functional"
+
+  _dkexec bash -c "cp -f \${DOC_ROOT}/sites/simpletest/browser_output/*.html \${REPORT_DIR}/functional"
 }
 
 _functional_js() {
@@ -208,9 +205,11 @@ _functional_js() {
 
   _dkexec bash -c "curl -s http://localhost:4444/wd/hub/status | jq '.'"
 
-  _dkexec robo test:suite "${PHPUNIT_TESTS}functional-javascript"
+  _dkexec "chown -R \${APACHE_RUN_USER}:\${APACHE_RUN_GROUP} \${WEB_ROOT}/sites/"
+  _dkexec "sudo -E -u \${APACHE_RUN_USER} robo test:suite \${PHPUNIT_TESTS}functional-javascript"
+  # _dkexec robo test:suite "\${PHPUNIT_TESTS}functional-javascript"
 
-  _dkexec bash -c "cp -f ${DOC_ROOT}/sites/simpletest/browser_output/*.html ${REPORT_DIR}/functional-javascript"
+  _dkexec bash -c "cp -f \${DOC_ROOT}/sites/simpletest/browser_output/*.html \${REPORT_DIR}/functional-javascript"
 }
 
 _nightwatch() {
@@ -219,20 +218,20 @@ _nightwatch() {
   _build
   _tests_prepare
 
-  _dkexec bash -c "cp -u ${CI_PROJECT_DIR}/.gitlab-ci/.env.nightwatch ${WEB_ROOT}/core/.env"
+  _dkexec bash -c "cp -u \${CI_PROJECT_DIR}/.gitlab-ci/.env.nightwatch \${WEB_ROOT}/core/.env"
 
   printf ">>> [NOTICE] Patching nightwatch for Drupal profile support.\\n"
 
-  _dkexec bash -c "curl -fsSL https://www.drupal.org/files/issues/2019-02-05/3017176-7.patch -o ${DOC_ROOT}/3017176-7.patch"
-  docker exec -it -w ${WEB_ROOT} ci-drupal bash -c "patch -N -p1 < ${DOC_ROOT}/3017176-7.patch"
+  _dkexec bash -c "curl -fsSL https://www.drupal.org/files/issues/2019-02-05/3017176-7.patch -o \${DOC_ROOT}/3017176-7.patch"
+  docker exec -it -w ${WEB_ROOT} ci-drupal bash -c "patch -N -p1 < \${DOC_ROOT}/3017176-7.patch"
 
-  _dkexec bash -c "cp ${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php ${CI_PROJECT_DIR}"
+  _dkexec bash -c "cp \${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php \${CI_PROJECT_DIR}"
   _dkexec robo prepare:folders
 
-  _dkexec chown -R ${APACHE_RUN_USER}:${APACHE_RUN_GROUP} ${DOC_ROOT}
+  _dkexec "chown -R \${APACHE_RUN_USER}:\${APACHE_RUN_GROUP} \${DOC_ROOT}"
 
-  docker exec -it -w ${WEB_ROOT}/core ci-drupal yarn install
-  docker exec -it -w ${WEB_ROOT}/core ci-drupal yarn test:nightwatch ${NIGHTWATCH_TESTS}
+  docker exec -it -w ${WEB_ROOT}/core ci-drupal "yarn install"
+  docker exec -it -w ${WEB_ROOT}/core ci-drupal "yarn test:nightwatch \${NIGHTWATCH_TESTS}"
 }
 
 _behat() {
@@ -241,14 +240,14 @@ _behat() {
   _build
   _tests_prepare
 
-  _dkexec bash -c "cp ${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php ${CI_PROJECT_DIR}"
+  _dkexec bash -c "cp \${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php \${CI_PROJECT_DIR}"
   _dkexec robo prepare:folders
   _dkexec robo install:drupal standard
 
-  _dkexec chown -R ${APACHE_RUN_USER}:${APACHE_RUN_GROUP} ${WEB_ROOT}/sites/
+  _dkexec "chown -R \${APACHE_RUN_USER}:\${APACHE_RUN_GROUP} \${WEB_ROOT}/sites/"
 
   # Starting Chrome.
-  docker exec -d ci-drupal bash -c "/usr/bin/chromium --no-sandbox --disable-gpu --headless --remote-debugging-address=0.0.0.0 --remote-debugging-port=9222 --window-size=1920,1080"
+  docker exec -d ci-drupal bash -c "/scripts/start-chrome.sh"
   _dkexec bash -c "curl -s http://localhost:9222/json/version | jq '.'"
 
   _dkexec robo test:behat
@@ -258,7 +257,7 @@ _security_checker() {
   printf "\\n%s[INFO]%s Perform job 'Security report' (security_checker)\\n\\n" "${_blu}" "${_end}"
 
   _build
-  _dkexec bash -c "cp ${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php ${CI_PROJECT_DIR}"
+  _dkexec bash -c "cp \${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php \${CI_PROJECT_DIR}"
   _dkexec robo prepare:folders
   _dkexec security-checker security:check
 }
@@ -268,8 +267,8 @@ _security_checker() {
 # Replicate cp in all qa / lint / metrics jobs
 _cp_qa_lint_metrics() {
   # Place config files in a proper directory.
-  printf ">>> [NOTICE] cp config to %s\\n" "${CI_PROJECT_DIR}"
-  _dkexec bash -c "cp ${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php ${CI_PROJECT_DIR}/.gitlab-ci/.phpmd.xml ${CI_PROJECT_DIR}/.gitlab-ci/.phpqa.yml ${CI_PROJECT_DIR}/.gitlab-ci/.eslintignore ${CI_PROJECT_DIR}"
+  printf ">>> [NOTICE] cp config\\n"
+  _dkexec bash -c "cp \${CI_PROJECT_DIR}/.gitlab-ci/RoboFile.php \${CI_PROJECT_DIR}/.gitlab-ci/.phpmd.xml \${CI_PROJECT_DIR}/.gitlab-ci/.phpqa.yml \${CI_PROJECT_DIR}/.gitlab-ci/.eslintignore \${CI_PROJECT_DIR}"
 }
 
 _code_quality() {
@@ -277,7 +276,7 @@ _code_quality() {
   _cp_qa_lint_metrics
   _dkexec robo prepare:folders
 
-  _dkexec phpqa ${PHPQA_REPORT}/code_quality ${TOOLS} ${PHPQA_PHP_CODE}
+  _dkexec "phpqa \${PHPQA_REPORT}/code_quality \${TOOLS} \${PHPQA_PHP_CODE}"
 }
 
 _best_practices() {
@@ -286,7 +285,7 @@ _best_practices() {
   sed -i 's/Drupal/DrupalPractice/g' .phpqa.yml
   _dkexec robo prepare:folders
 
-  _dkexec phpqa ${PHPQA_REPORT}/best_practices --tools ${BEST_PRACTICES} ${PHPQA_PHP_CODE}
+  _dkexec "phpqa \${PHPQA_REPORT}/best_practices --tools \${BEST_PRACTICES} \${PHPQA_PHP_CODE}"
 }
 
 ####### Lint jobs
@@ -296,8 +295,8 @@ _eslint() {
   _cp_qa_lint_metrics
   _dkexec robo prepare:folders
 
-  # _dkexec eslint --config ${WEB_ROOT}/core/.eslintrc.passing.json ${JS_CODE}
-  _dkexec eslint --config ${WEB_ROOT}/core/.eslintrc.passing.json --format html --output-file ${REPORT_DIR}/js-lint-report.html ${JS_CODE}
+  # _dkexec "eslint --config \${WEB_ROOT}/core/.eslintrc.passing.json \${JS_CODE}"
+  _dkexec "eslint --config \${WEB_ROOT}/core/.eslintrc.passing.json --format html --output-file \${REPORT_DIR}/js-lint-report.html \${JS_CODE}"
 }
 
 _stylelint() {
@@ -305,8 +304,8 @@ _stylelint() {
   _cp_qa_lint_metrics
   _dkexec robo prepare:folders
 
-  _dkexec stylelint --config-basedir /var/www/.node/node_modules/ \
-    --config ${WEB_ROOT}/core/.stylelintrc.json -f verbose "${CSS_FILES}"
+  _dkexec "stylelint --config-basedir /var/www/.node/node_modules/ \
+    --config \${WEB_ROOT}/core/.stylelintrc.json -f verbose \${CSS_FILES}"
 }
 
 _sass_lint() {
@@ -317,8 +316,8 @@ _sass_lint() {
   printf ">>> [NOTICE] Install Sass-lint\\n"
   docker exec -it -w ${WEB_ROOT}/core ci-drupal npm install --no-audit git://github.com/sasstools/sass-lint.git#develop
 
-  _dkexec ${WEB_ROOT}/core/node_modules/.bin/sass-lint --config ${SASS_CONFIG} --verbose --no-exit
-  _dkexec ${WEB_ROOT}/core/node_modules/.bin/sass-lint --config ${SASS_CONFIG} --verbose --no-exit --format html --output ${REPORT_DIR}/sass-lint-report.html
+  _dkexec "\${WEB_ROOT}/core/node_modules/.bin/sass-lint --config \${SASS_CONFIG} --verbose --no-exit"
+  _dkexec "\${WEB_ROOT}/core/node_modules/.bin/sass-lint --config \${SASS_CONFIG} --verbose --no-exit --format html --output \${REPORT_DIR}/sass-lint-report.html"
 }
 
 ####### Metrics jobs
@@ -328,7 +327,7 @@ _phpmetrics() {
   _cp_qa_lint_metrics
   _dkexec robo prepare:folder
 
-  _dkexec phpqa ${PHPQA_REPORT}/phpmetrics --tools phpmetrics ${PHPQA_PHP_CODE}
+  _dkexec "phpqa \${PHPQA_REPORT}/phpmetrics --tools phpmetrics \${PHPQA_PHP_CODE}"
 }
 
 _phpstat() {
@@ -336,7 +335,7 @@ _phpstat() {
   _cp_qa_lint_metrics
   _dkexec robo prepare:folders
 
-  _dkexec phpqa ${PHPQA_REPORT}/phpstat --tools phploc,pdepend ${PHPQA_PHP_CODE}
+  _dkexec "phpqa \${PHPQA_REPORT}/phpstat --tools phploc,pdepend \${PHPQA_PHP_CODE}"
 }
 
 ###############################################################################
@@ -359,6 +358,61 @@ _dkexeca() {
 # Helpers commands.
 ###############################################################################
 
+_init_variables() {
+  CI_TYPE=$(yq r ./.gitlab-ci.yml variables.CI_TYPE)
+  WEB_ROOT=$(yq r ./.gitlab-ci.yml variables.WEB_ROOT)
+  CI_PROJECT_DIR="/builds"
+  printf "CI_TYPE: %s\\nWEB_ROOT: %s\\nCI_PROJECT_DIR: %s\\n\\n" ${CI_TYPE} ${WEB_ROOT} ${CI_PROJECT_DIR}
+}
+
+_generate_env_from_yaml() {
+  printf "\\n>>> [NOTICE] Generate .env file..."
+
+  if ! [ -x "$(command -v yq)" ]; then
+    curl -fsSL https://github.com/mikefarah/yq/releases/download/2.4.0/yq_linux_amd64 -o /usr/local/bin/yq && chmod +x /usr/local/bin/yq
+  fi
+
+  if ! [ -f "./.gitlab-ci.yml" ]; then
+    printf "%s[ERROR]%s Missing .gitlab-ci.yml file.\\n" "${_red}" "${_end}"
+    exit 1
+  fi
+  __yaml="./.gitlab-ci.yml"
+  __env="./local/.gitlab-ci.env"
+
+  if [ -f $__env ]; then
+    rm -f $__env
+  fi
+
+  touch $__env
+  echo 'CI_PROJECT_NAME: my_module' >> $__env
+  echo 'CI_PROJECT_DIR: /builds' >> $__env
+
+  yq r $__yaml variables >> $__env
+  yq r $__yaml "[.test_variables]" >> $__env
+  # yq r $__yaml "Behat tests.variables" >> $__env
+
+  sed -i "s#\${CI_PROJECT_DIR}#/builds#g" $__env
+
+  __REPORT_DIR=$(yq r $__yaml variables.REPORT_DIR)
+  sed -i "s#\${REPORT_DIR}#$__REPORT_DIR#g" $__env
+
+  __PHP_CODE=$(yq r $__yaml variables.PHP_CODE)
+  sed -i "s#\${PHP_CODE}#$__PHP_CODE#g" $__env
+
+  __PHPQA_IGNORE_DIRS=$(yq r $__yaml variables.PHPQA_IGNORE_DIRS)
+  sed -i "s#\${PHPQA_IGNORE_DIRS}#$__PHPQA_IGNORE_DIRS#g" $__env
+
+  __PHPQA_IGNORE_FILES=$(yq r $__yaml variables.PHPQA_IGNORE_FILES)
+  sed -i "s#\${PHPQA_IGNORE_FILES}#$__PHPQA_IGNORE_FILES#g" $__env
+
+  sed -i 's#: #=#g' $__env
+
+  __WEB_ROOT=$(yq r $__yaml variables.WEB_ROOT)
+  sed -i "s#\${WEB_ROOT}#$__WEB_ROOT#g" $__env
+
+  printf ">>> %s ... Done!\\n" $__env
+}
+
 _reset_gitlab() {
   printf "\\n%s[INFO]%s Reset stack to mimic Gitlab-ci\\n" "${_blu}" "${_end}"
   _down
@@ -374,6 +428,10 @@ _nuke() {
 }
 
 _up() {
+  if ! [ -f "./local/.gitlab-ci.env" ]; then
+    _generate_env_from_yaml
+  fi
+
   if [ -f "local/docker-compose.yml" ]; then
     docker-compose -f local/docker-compose.yml up -d
   else
@@ -400,9 +458,14 @@ _down() {
   fi
 }
 
+_copy_output() {
+  _dkexec bash -c "cp -f ${DOC_ROOT}/sites/simpletest/browser_output/*.html ${REPORT_DIR}/"
+}
+
 _clean() {
   _clean_config
-  rm -rf reports/*
+  sudo rm -rf reports/*
+  _dkexec bash -c "rm -rf ${DOC_ROOT}/sites/simpletest/browser_output/*.html"
 }
 
 _clean_config() {
@@ -488,13 +551,18 @@ _main() {
   if [ "${_CMD}" == "help" ]; then
     _help
     exit 0
+  elif [ "${_CMD}" == "init_variables" ] || [ "${_CMD}" == "init" ]; then
+    _init_variables
+    exit 0
   fi
 
   # Run command if exist.
   __call="_${_CMD}"
   if [ "$(type -t "${__call}")" == 'function' ]; then
+    _init_variables
     $__call
   else
+    printf "%s[ERROR]%s Unknown command: %s\\n" "${_red}" "${_end}" "${_CMD}"
     _help
   fi
 }
